pyproject.toml
[project]
name = "project-julep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "julep>=2.10.0",
    "langchain>=0.3.25",
    "langgraph>=0.4.8",
    "numpy>=2.2.6",
    "python-dotenv>=1.0.1",
    "streamlit>=1.45.1",
    "uuid>=1.30",
]
